requirements.txt
web3==7.0.0b2
discord.py
python-dotenv
